+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - Unreleased
+### Added
+- New attribute `align` to `ola-panel-text` to align the text horizontally
+
+### Changed
+- `ola-counter` component structure. Now the component includes the title and description, not only the number.
+
+### Fixed
+- Default styles of `<ul>`
+
 ## [0.1.3] - 2020-07-08
 ### Changed
 - Appearance of `ola-counter`: now it's smaller and with a shadow
@@ -29,6 +39,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.1.0 - 2020-04-23
 First version
 
+[0.2.0]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.3...HEAD
 [0.1.3]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/marketgoo/Ola-Emails/compare/v0.1.0...v0.1.1